pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deeprecon"
version = "1.0.0"
authors = [
    {name = "Mohammad Rasol Esfandiari", email = "mrasolesfandiari@gmail.com"},
]
description = "A powerful, modular Python library for comprehensive domain and IP analysis"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Security",
    "Topic :: System :: Networking",
]
keywords = [
    "domain", "ip", "dns", "whois", "ssl", "security",
    "networking", "reconnaissance", "analysis", "geolocation"
]
dependencies = [
    "requests>=2.31.0",
    "dnspython>=2.4.2",
    "python-whois>=0.8.0",
    "ipwhois>=1.2.0",
    "pyOpenSSL>=23.3.0",
    "beautifulsoup4>=4.12.2",
    "lxml>=4.9.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "flake8>=5.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
]

[project.urls]
Homepage = "https://github.com/DeepPythonist/DeepRecon"
Repository = "https://github.com/DeepPythonist/DeepRecon"
Documentation = "https://github.com/DeepPythonist/DeepRecon#readme"
Changelog = "https://github.com/DeepPythonist/DeepRecon/blob/main/CHANGELOG.md"
"Bug Reports" = "https://github.com/DeepPythonist/DeepRecon/issues"

[project.scripts]
deeprecon = "cli:main"

[tool.setuptools]
packages = ["deeprecon"]
include-package-data = true

[tool.setuptools.package-data]
deeprecon = ["locales/*.json"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["deeprecon"]

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [".git", "__pycache__", "build", "dist"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"] 